Moved logic from event repo to service
