+++ HobbyMatch.Database/Repositories/Events/IEventRepository.cs
@@ -5,9 +5,8 @@ namespace HobbyMatch.Database.Repositories.Events
     public interface IEventRepository
     {
         Task<Event?> GetEventByIdAsync(int eventId);
-        Task<bool> AddUserToEventAsync(int eventId, User user);
         Task<Event?> AddEvent(Event newEvent);
-		Task<bool> RemoveUserFromEventAsync(int eventId, User user);
         Task<List<Event>> GetEventsWithFilter(string? filter);
-    }
+		Task SaveChangesAsync();
+	}
 }
